CC2538: add secondary location to ieee address
The CC2538 currently has two addressing options: a hardcoded address set
at compile time or the address stored in primary address section of the
info page. This commit adds the option to choose the secondary location
of the ieee address from the info page, or any memory address.

To use, define `IEEE_ADDR_CONF_LOCATION` in `project-conf.h` or similar.

For example:

    #define IEEE_ADDR_CONF_LOCATION IEEE_ADDR_LOCATION_SECONDARY